Add SignupEntry component and update next_page routing
